struct default/Constant support
for struct inputs on library function nodes, if a struct has a default constructor, an input with a Constant will now be created. This allows the struct to be unwired and the default value used. The struct is not serialized though, because of new ConstantIsTransient input flag. Note this is lower priority than types registered with DefaultTypeInfoLibrary, which may still be serializable.

Note that this introduces a backcompat situation, as if the default struct changes, potentially so does the input pin value.